backend/requirements.txt
langchain
langchain-community
openai
langchain-openai
chromadb
python-dotenv
fastapi
uvicorn
tiktoken
pypdf
sentence-transformers